requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pyyaml>=6.0
scipy>=1.11.0
python-dateutil>=2.8.2
pytz>=2023.3
